requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlalchemy==2.0.31
typing-extensions==4.12.2
werkzeug==3.0.3
